requirements.txt
google-generativeai
numpy
supabase
